+++ QS.ServiceLayer/User/Dtos/InputDto/UserUpdateInputDto.cs
@@ -42,7 +42,7 @@ namespace QS.ServiceLayer.User.Dtos.InputDto
         /// <summary>
         /// 角色
         /// </summary>
-        public long[] RoleIds { get; set; }
+        public int[] RoleIds { get; set; }
 
         /// <summary>
         /// 版本
